% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fdr.R
\name{calculate_fdr}
\alias{calculate_fdr}
\title{Calculate FDR}
\usage{
calculate_fdr(average_points, extrema)
}
\arguments{
\item{average_points}{A `data.frame`. The output of [get_average_points_against()].}

\item{extrema}{A `data.frame`. The output of [get_extrema_points_against()].}
}
\value{
A `data.frame` containing the opponent; the location of the scoring team; the player position; the fixture difficulty
for the scoring player against the opponent.
}
\description{
Calculate FDR
}
