% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contender_functions.R
\name{contender_info_detailed}
\alias{contender_info_detailed}
\title{Fetch data for specific contender given gameweek}
\usage{
contender_info_detailed(
  id_contender = c(2045023),
  id_gameweek = c(1),
  link = "https://fantasy.premierleague.com/api/entry/"
)
}
\description{
Fetch data for specific contender given gameweek
}
