% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contender_functions.R
\name{contender_info_detailed_picks}
\alias{contender_info_detailed_picks}
\title{Fetch data for specific contenders pick a given gameweek}
\usage{
contender_info_detailed_picks(contender_info = contender_info_detailed())
}
\description{
Fetch data for specific contenders pick a given gameweek
}
