% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contender_functions.R
\name{contender_info_detailed_status}
\alias{contender_info_detailed_status}
\title{Fetch status for specific contenders a given gameweek}
\usage{
contender_info_detailed_status(contender_info = contender_info_detailed())
}
\description{
Fetch status for specific contenders a given gameweek
}
