% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/curl_async.R
\name{curl_async}
\alias{curl_async}
\title{Download Data}
\usage{
curl_async(urls, .parse = RcppSimdJson::fparse)
}
\arguments{
\item{urls}{`character(n)`. A vector of URLs.}

\item{.parse}{`function`. The function used to parse the JSON data.}
}
\value{
The parsed JSON.
}
\description{
Fetch multiple data sources at once and parse them once they are downloaded.
}
\examples{
curl_async("https://fantasy.premierleague.com/api/element-summary/1/")

}
