% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/web.R
\name{filter_players}
\alias{filter_players}
\title{Filter player elements from the web}
\usage{
filter_players(data)
}
\arguments{
\item{data}{list from api}
}
\value{
returns data frame with additional columns, player photo paths, and team abbreviations
}
\description{
Filter player elements from the web
}
\examples{
\dontrun{
data_list <- get_current_stats()
players <- filter_players(data_list)
}
}
