% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fixtures_functions.R
\name{fixtures_functions}
\alias{fixtures_functions}
\title{Selecting fixtures from API}
\usage{
fixtures_functions(
  upcoming_fixtures = upcoming_fixtures,
  team_data = players_teams_data \%>\% team_functions(.)
)
}
\description{
Selecting fixtures from API
}
