% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fpl-package.R
\docType{package}
\name{fpl-package}
\alias{fpl}
\alias{fpl-package}
\title{fpl}
\description{
fpl
}
\details{
fpl provides access to the Barclays Fantasy Premier League API
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/nathaneastwood/fpl}
  \item Report bugs at \url{https://github.com/nathaneastwood/fpl/issues}
}

}
\author{
\strong{Maintainer}: Nathan Eastwood \email{nathan.eastwood@icloud.com}

}
