% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fpl.R
\name{FPL}
\alias{FPL}
\title{R6 Class representing the Fantasy Premier League's API}
\description{
The FPL class is the main class used for interacting with Fantasy Premier League’s API.
}
\details{
Information is taken from <https://fantasy.premierleague.com/api/bootstrap-static/>.
}
\examples{

## ------------------------------------------------
## Method `FPL$new`
## ------------------------------------------------

fpl <- FPL$new()

## ------------------------------------------------
## Method `FPL$get_user`
## ------------------------------------------------

fpl$get_user(555690)


## ------------------------------------------------
## Method `FPL$get_team`
## ------------------------------------------------

fpl$get_team(1)


## ------------------------------------------------
## Method `FPL$get_teams`
## ------------------------------------------------

fpl$get_teams(1)


## ------------------------------------------------
## Method `FPL$get_player`
## ------------------------------------------------

fpl$get_player(1)


## ------------------------------------------------
## Method `FPL$get_players`
## ------------------------------------------------

fpl$get_players(c(1, 10, 100))


## ------------------------------------------------
## Method `FPL$get_player_summary`
## ------------------------------------------------

fpl$get_player_summary(1)


## ------------------------------------------------
## Method `FPL$get_player_summaries`
## ------------------------------------------------

fpl$get_player_summaries(c(1, 10, 100))

}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{events}}{A `data.frame`. Game week event data.}

\item{\code{game_settings}}{A `list`. FPL game settings.}

\item{\code{phases}}{A `data.frame`. Monthly gameweek information.}

\item{\code{teams}}{A `data.frame`. Information pertaining to a Premier League team.}

\item{\code{total_players}}{`integer(1)`. The total number of participants in FPL.}

\item{\code{elements}}{A `data.frame`. The players data.}

\item{\code{element_stats}}{A `data.frame`. Labels for certain stats based columns.}

\item{\code{element_types}}{A `data.frame`. Squad position information.}
}
\if{html}{\out{</div>}}
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{current_gameweek}}{`integer(1)`. The current gameweek.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{FPL$new()}}
\item \href{#method-get_user}{\code{FPL$get_user()}}
\item \href{#method-get_team}{\code{FPL$get_team()}}
\item \href{#method-get_teams}{\code{FPL$get_teams()}}
\item \href{#method-get_player}{\code{FPL$get_player()}}
\item \href{#method-get_players}{\code{FPL$get_players()}}
\item \href{#method-get_player_summary}{\code{FPL$get_player_summary()}}
\item \href{#method-get_player_summaries}{\code{FPL$get_player_summaries()}}
\item \href{#method-clone}{\code{FPL$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Instantiate the `FPL` class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FPL$new()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
An R6 `FPL` object.
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{fpl <- FPL$new()
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_user"></a>}}
\if{latex}{\out{\hypertarget{method-get_user}{}}}
\subsection{Method \code{get_user()}}{
Get information about a specific user.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FPL$get_user(user_id)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{user_id}}{`numeric(1)`. The user's id.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
An R6 [User()] object.
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{fpl$get_user(555690)

}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_team"></a>}}
\if{latex}{\out{\hypertarget{method-get_team}{}}}
\subsection{Method \code{get_team()}}{
Get information about a specific team.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FPL$get_team(team_id)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{team_id}}{`numeric(1)`. The team's id.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
An R6 `Team` object.
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{fpl$get_team(1)

}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_teams"></a>}}
\if{latex}{\out{\hypertarget{method-get_teams}{}}}
\subsection{Method \code{get_teams()}}{
Get information about specific teams.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FPL$get_teams(team_ids = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{team_id}}{`numeric(n)`. The team ids.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A `list` of R6 [Team()] objects.
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{fpl$get_teams(1)

}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_player"></a>}}
\if{latex}{\out{\hypertarget{method-get_player}{}}}
\subsection{Method \code{get_player()}}{
Get information about a specific player.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FPL$get_player(player_id, include_summary = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{player_id}}{`numeric(1)`. The player's id.}

\item{\code{include_summary}}{`logical(1)`. Whether to include historical data and fixture details for the player.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
An R6 [Player()] object.
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{fpl$get_player(1)

}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_players"></a>}}
\if{latex}{\out{\hypertarget{method-get_players}{}}}
\subsection{Method \code{get_players()}}{
Get information about multiple players.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FPL$get_players(player_ids = NULL, include_summary = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{player_ids}}{`numeric(n)`. The player ids.}

\item{\code{include_summary}}{`logical(1)`. Whether to include historical data and fixture details for the player.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A `list` of R6 [Player()] objects.
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{fpl$get_players(c(1, 10, 100))

}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_player_summary"></a>}}
\if{latex}{\out{\hypertarget{method-get_player_summary}{}}}
\subsection{Method \code{get_player_summary()}}{
Get historical and fixture information about a specific player.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FPL$get_player_summary(player_id)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{player_id}}{`numeric(1)`. The player's id.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
An R6 [PlayerSummary()] object.
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{fpl$get_player_summary(1)

}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_player_summaries"></a>}}
\if{latex}{\out{\hypertarget{method-get_player_summaries}{}}}
\subsection{Method \code{get_player_summaries()}}{
Get historical and fixture information about multiple players.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FPL$get_player_summaries(player_id)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{player_id}}{`numeric(n)`. The player ids.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A `list` of R6 [PlayerSummary()] objects.
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{fpl$get_player_summaries(c(1, 10, 100))

}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FPL$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
