% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{fpl_plot}
\alias{fpl_plot}
\title{Plot two variables with a linear square regression}
\usage{
fpl_plot(
  spreadsheet,
  x = "cost",
  y = "points_per_game",
  group = "names",
  pos = NULL,
  top_percent = NULL,
  remove_player = NULL
)
}
\arguments{
\item{spreadsheet}{input data frame, spreadsheet, or path to data}

\item{x}{default: "cost", name of column in spreadsheet variable. Input is flexible
and will return the shortest partial match. E.g., with columns: "cost", "starting_cost",
the "cost" column will be selected because it is shortest.}

\item{y}{default: "points_per_game", name of column in spreadsheet variable,
similar to x variable. Will by plotted on the y-axis}

\item{group}{default: "names". Group or column to plot categories by.
Options:
"names" - plots points as players display names
"team" - plots points as team icon
"images" - plots points as player images}

\item{pos}{default: NULL. Option to filter players by position.
Options:
"GKP" - Goalies,
"FWD" - Forwards,
"MID" - Midfielders,
"DEF" - Defenders}

\item{top_percent}{default: NULL. Option to filter players by total points
percentage. Expects values from 0 - 1.}

\item{remove_player}{default: NULL. Input vector of player(s) to remove players from
plotting or regression. E.g. the input c("Haaland") will remove him from the table.}
}
\value{
Plot of input parameters with linear square regression and regression
coefficients. Additionally, prints out filtered table.
}
\description{
Data or spreadsheet can be obtained from the premier league API. Using
get_player_stats() followed by filter_players() will obtain the current player
records.
}
\examples{
\dontrun{
stats <- get_premier_stats()
spreadsheet <- filter_players(stats)
fpl_plot(spreadsheet, x = "cost", y = "total_points", group = "names", pos = "FWD")
}
}
