% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gameweek_functions.R
\name{gameweek_functions}
\alias{gameweek_functions}
\title{Selecting team data from API}
\usage{
gameweek_functions(
  players_teams_data = players_teams_data,
  link_users = "https://fantasy.premierleague.com/entry/"
)
}
\description{
Selecting team data from API
}
