% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fdr.R
\name{get_average_points_against}
\alias{get_average_points_against}
\title{Get Average Points Against}
\usage{
get_average_points_against(points_against)
}
\arguments{
\item{points_against}{A `data.frame` of points scored against all teams in the Premier League, typically the output
of [FPL$get_points_against()].}
}
\value{
A `data.frame` containing the `opponent`; fixture `location`; player `position`; and the `points` scored.
}
\description{
Get Average Points Against
}
