% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/web.R
\name{get_current_stats}
\alias{get_current_stats}
\title{Get the current the stats from the fantasy API}
\usage{
get_current_stats()
}
\value{
data frame with all players and variables
}
\description{
Get the current the stats from the fantasy API
}
\examples{
# Get the stats from online
weekly_stats <- get_current_stats()
head(weekly_stats) # print stats

}
