% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fdr.R
\name{get_extrema_points_against}
\alias{get_extrema_points_against}
\title{Get Extrema Points Against}
\usage{
get_extrema_points_against(points_against)
}
\arguments{
\item{points_against}{A `data.frame` of points scored against all teams in the Premier League, typically the output
of [FPL$get_points_against()].}
}
\value{
A `data.frame` containing fixture `location`; player `position`; and the `min`imum and `max`imum points against.
}
\description{
Get Extrema Points Against
}
