% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_transform_data.R
\name{load_data}
\alias{load_data}
\title{Fetch high-level data for players, teams and fixtures}
\usage{
load_data(
  players_teams_data_link = "https://fantasy.premierleague.com/api/bootstrap-static/",
  upcoming_fixtures_link = "https://fantasy.premierleague.com/api/fixtures"
)
}
\description{
Fetch high-level data for players, teams and fixtures
}
