% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/player.R
\name{Player}
\alias{Player}
\title{R6 Class Representing a Player}
\description{
The [FPL$get_player()] and [FPL$get_players()] methods return objects of class `Player`.
}
\details{
Information is taken from <https://fantasy.premierleague.com/api/bootstrap-static/>.
}
\examples{

## ------------------------------------------------
## Method `Player$new`
## ------------------------------------------------

fpl <- FPL$new()
player <- fpl$get_player(250)
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{player}}{A `data.frame`. Player data for a single player.}

\item{\code{history}}{A `data.frame`. The current season's data for a given player. Empty by default unless particular
active bindings are called.}
}
\if{html}{\out{</div>}}
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{games_played}}{`integer(1)`.}

\item{\code{price}}{`numeric(1)`. The player's in game price.}

\item{\code{pp90}}{`numeric(1)`. Points scored per 90 minutes.}

\item{\code{ppm}}{`numeric(1)`. Points scored per match.}

\item{\code{ppmm}}{`numeric(1)`. Points scored per match per £1M.}

\item{\code{vapm}}{`numeric(1)`. Value added per £1M.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{Player$new()}}
\item \href{#method-print}{\code{Player$print()}}
\item \href{#method-clone}{\code{Player$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Instantiate the `Player` class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Player$new(player)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{player}}{A `data.frame`. Player data for a single player.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
An R6 object of class `Player`.
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{fpl <- FPL$new()
player <- fpl$get_player(250)
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-print"></a>}}
\if{latex}{\out{\hypertarget{method-print}{}}}
\subsection{Method \code{print()}}{
Pretty printing of the the `Player`.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Player$print(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{Not used.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Player$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
