% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/player_functions.R
\name{player_functions}
\alias{player_functions}
\title{General player information (aggregated)}
\usage{
player_functions(players_teams_data = players_teams_data)
}
\description{
General player information (aggregated)
}
