% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/player_functions.R
\name{player_info_detailed}
\alias{player_info_detailed}
\title{Fetch data for specific player}
\usage{
player_info_detailed(
  id = c(5, 50),
  link = "https://fantasy.premierleague.com/api/element-summary/"
)
}
\description{
Fetch data for specific player
}
