% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/player_functions.R
\name{player_info_detailed_fixtures_current}
\alias{player_info_detailed_fixtures_current}
\title{Fetch data for specific player current season}
\usage{
player_info_detailed_fixtures_current(player_info = player_info_detailed())
}
\description{
Fetch data for specific player current season
}
