% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/player_functions.R
\name{player_info_detailed_fixtures_previous}
\alias{player_info_detailed_fixtures_previous}
\title{Fetch data for specific player previous seasons}
\usage{
player_info_detailed_fixtures_previous(player_info = player_info_detailed())
}
\description{
Fetch data for specific player previous seasons
}
