% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/player_functions.R
\name{player_info_detailed_fixtures_upcoming}
\alias{player_info_detailed_fixtures_upcoming}
\title{Fetch data for specific player - upcoming fixtures}
\usage{
player_info_detailed_fixtures_upcoming(player_info = player_info_detailed())
}
\description{
Fetch data for specific player - upcoming fixtures
}
