% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/player.R
\name{PlayerSummary}
\alias{PlayerSummary}
\title{R6 Class Representing a Player's Historic Data}
\description{
The [FPL$get_player_summary()] and [FPL$get_player_summaries()] methods return objects of class `PlayerSummary`.
}
\details{
Information is taken from <https://fantasy.premierleague.com/api/element-summary/{player_id}/>.
}
\examples{

## ------------------------------------------------
## Method `PlayerSummary$new`
## ------------------------------------------------

fpl <- FPL$new()
player_summary <- fpl$get_player_summary(250)
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{fixtures}}{A `data.frame`. The player's fixture list for the season.}

\item{\code{history}}{A `data.frame`. The player's data for the current seasom.}

\item{\code{history_past}}{A `data.frame`. The player's summary data for previous seasons.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{PlayerSummary$new()}}
\item \href{#method-clone}{\code{PlayerSummary$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Instantiate the `PlayerSummary` class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PlayerSummary$new(player_summary)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{player_summary}}{`numeric(1)`. The player's id.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
An R6 object of class `PlayerSummary`.
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{fpl <- FPL$new()
player_summary <- fpl$get_player_summary(250)
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PlayerSummary$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
