% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/converters.R
\name{position_converter}
\alias{position_converter}
\title{Position converter}
\usage{
position_converter(position)
}
\arguments{
\item{position}{`numeric(1)`. The player's position.}
}
\value{
A `character` string of the player's actual position.
}
\description{
Converts a player's `element_type` to their actual position.
}
