% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/team.R
\name{Team}
\alias{Team}
\title{R6 Class Representing a Team}
\description{
The [FPL$get_team()] and [FPL$get_teams()] methods return objects of class `Team`.
}
\details{
Information is taken from <https://fantasy.premierleague.com/api/bootstrap-static/>.
}
\examples{

## ------------------------------------------------
## Method `Team$new`
## ------------------------------------------------

fpl <- FPL$new()
team <- fpl$get_team(11)

## ------------------------------------------------
## Method `Team$get_players`
## ------------------------------------------------

team$get_players()

}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{team}}{A `data.frame`. Data for a single Premier League team.}

\item{\code{players}}{A `data.frame`. Player data for a single Premier League team. Empty by default unless the
[FPL$get_players()] method is called.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{Team$new()}}
\item \href{#method-get_players}{\code{Team$get_players()}}
\item \href{#method-print}{\code{Team$print()}}
\item \href{#method-clone}{\code{Team$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Instantiate the `Team` class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Team$new(team)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{team}}{A `data.frame`. Data for a single Premier League team.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
An R6 object of class `Team`.
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{fpl <- FPL$new()
team <- fpl$get_team(11)
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_players"></a>}}
\if{latex}{\out{\hypertarget{method-get_players}{}}}
\subsection{Method \code{get_players()}}{
Get information about specific players from the `team`.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Team$get_players()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A `list` of R6 `Player` objects who play for the `Team`.
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{team$get_players()

}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-print"></a>}}
\if{latex}{\out{\hypertarget{method-print}{}}}
\subsection{Method \code{print()}}{
Pretty printing of the `Team`.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Team$print(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{Not used.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Team$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
