% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/converters.R
\name{team_converter}
\alias{team_converter}
\title{Team Converter}
\usage{
team_converter(teams, team_id)
}
\arguments{
\item{teams}{A `data.frame` containing the teams data. See details for more.}

\item{team_id}{`numeric(1)`. The team's id.}
}
\value{
A `character(1)` containing the team's actual name.
}
\description{
Convert a team's ID to their actual name.
}
\details{
`teams` is typically the output of the teams element of `https://fantasy.premierleague.com/api/bootstrap-static/`.
}
