% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{test_logistic_count}
\alias{test_logistic_count}
\title{Test error for logistic regression with counting error}
\usage{
test_logistic_count(X_train, y_train, X_test, y_test)
}
\arguments{
\item{X_train}{a n x p matrix of training regressor}

\item{y_train}{a n-vector of training response}

\item{X_test}{a m x p matrix of test regressor}

\item{y_test}{a m-vector of test response}
}
\description{
Test error for logistic regression with counting error
}
