% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_transform_data.R
\name{transform_data}
\alias{transform_data}
\title{Transform high-level data for players, teams and fixtures}
\usage{
transform_data(
  loaded_data = load_data(),
  link_users = "https://fantasy.premierleague.com/entry/"
)
}
\description{
Transform high-level data for players, teams and fixtures
}
