% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/user.R
\name{User}
\alias{User}
\title{R6 Class representing a User}
\description{
The [FPL$get_user()] method returns an object of class `User`.
}
\examples{

## ------------------------------------------------
## Method `User$new`
## ------------------------------------------------

fpl <- FPL$new()
user <- fpl$get_user(555690)

## ------------------------------------------------
## Method `User$get_gameweek_history`
## ------------------------------------------------

user$get_gameweek_history()

## ------------------------------------------------
## Method `User$get_season_history`
## ------------------------------------------------

user$get_season_history()

## ------------------------------------------------
## Method `User$get_chips_history`
## ------------------------------------------------

user$get_gameweek_history()

## ------------------------------------------------
## Method `User$get_picks`
## ------------------------------------------------

user$get_picks()

## ------------------------------------------------
## Method `User$get_cup_status`
## ------------------------------------------------

user$get_cup_status()

## ------------------------------------------------
## Method `User$get_cup_matches`
## ------------------------------------------------

user$get_cup_matches()

## ------------------------------------------------
## Method `User$get_active_chips`
## ------------------------------------------------

user$get_active_chips()

## ------------------------------------------------
## Method `User$get_automatic_substitutions`
## ------------------------------------------------

user$get_automatic_substitutions()

## ------------------------------------------------
## Method `User$get_user_history`
## ------------------------------------------------

user$get_user_history()

## ------------------------------------------------
## Method `User$get_transfers`
## ------------------------------------------------

user$get_transfers()
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{user}}{A `list` of user details.}

\item{\code{history}}{A `list` of `data.frame`s containing the historic gameweek performance data for the user.}

\item{\code{cup}}{A `list` of data related to the user's cup performance.}

\item{\code{picks}}{A `list` of the user's picks information for each gameweek.}

\item{\code{transfers}}{A `list` of the user's weekly transfers.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{User$new()}}
\item \href{#method-get_gameweek_history}{\code{User$get_gameweek_history()}}
\item \href{#method-get_season_history}{\code{User$get_season_history()}}
\item \href{#method-get_chips_history}{\code{User$get_chips_history()}}
\item \href{#method-get_picks}{\code{User$get_picks()}}
\item \href{#method-get_cup_status}{\code{User$get_cup_status()}}
\item \href{#method-get_cup_matches}{\code{User$get_cup_matches()}}
\item \href{#method-get_active_chips}{\code{User$get_active_chips()}}
\item \href{#method-get_automatic_substitutions}{\code{User$get_automatic_substitutions()}}
\item \href{#method-get_user_history}{\code{User$get_user_history()}}
\item \href{#method-get_transfers}{\code{User$get_transfers()}}
\item \href{#method-print}{\code{User$print()}}
\item \href{#method-clone}{\code{User$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Instantiate the `User` class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{User$new(user)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{user}}{`list`. Data for a single FPL user.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
An R6 `User` object.
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{fpl <- FPL$new()
user <- fpl$get_user(555690)
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_gameweek_history"></a>}}
\if{latex}{\out{\hypertarget{method-get_gameweek_history}{}}}
\subsection{Method \code{get_gameweek_history()}}{
Get the gameweek history for the user.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{User$get_gameweek_history(gameweek = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{gameweek}}{`numeric(n)`. The gameweek history to return (default: `NULL` returns all gameweeks).}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A `data.frame`.
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{user$get_gameweek_history()
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_season_history"></a>}}
\if{latex}{\out{\hypertarget{method-get_season_history}{}}}
\subsection{Method \code{get_season_history()}}{
Get the seasonal history for the user.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{User$get_season_history()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A `data.frame`.
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{user$get_season_history()
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_chips_history"></a>}}
\if{latex}{\out{\hypertarget{method-get_chips_history}{}}}
\subsection{Method \code{get_chips_history()}}{
Get the chip history for the user.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{User$get_chips_history(gameweek = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{gameweek}}{`numeric(n)`. The gameweek history to return (default: `NULL` returns all gameweeks).}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A `data.frame`.
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{user$get_gameweek_history()
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_picks"></a>}}
\if{latex}{\out{\hypertarget{method-get_picks}{}}}
\subsection{Method \code{get_picks()}}{
Get the user's picks for the chosen `gameweek`(s).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{User$get_picks(gameweek = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{gameweek}}{`numeric(n)`. The gameweek picks to return (default: `NULL` returns all gameweeks).}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A `list` containing the user's picks details for the chosen gameweek(s).
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{user$get_picks()
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_cup_status"></a>}}
\if{latex}{\out{\hypertarget{method-get_cup_status}{}}}
\subsection{Method \code{get_cup_status()}}{
Get the user's cup status.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{User$get_cup_status()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A `list` of cup status information.
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{user$get_cup_status()
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_cup_matches"></a>}}
\if{latex}{\out{\hypertarget{method-get_cup_matches}{}}}
\subsection{Method \code{get_cup_matches()}}{
Get the user's cup matches.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{User$get_cup_matches(gameweek = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{gameweek}}{`numeric(n)`. The gameweek history to return (default: `NULL` returns all gameweeks).}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A `list` of cup match information.
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{user$get_cup_matches()
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_active_chips"></a>}}
\if{latex}{\out{\hypertarget{method-get_active_chips}{}}}
\subsection{Method \code{get_active_chips()}}{
Get the user's active chip for each gameweek or the active chip of the chosen `gameweek`(s).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{User$get_active_chips(gameweek = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{gameweek}}{`numeric(n)`. The gameweek history to return (default: `NULL` returns all gameweeks).}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A vector of active chips.
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{user$get_active_chips()
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_automatic_substitutions"></a>}}
\if{latex}{\out{\hypertarget{method-get_automatic_substitutions}{}}}
\subsection{Method \code{get_automatic_substitutions()}}{
Get a `list` of automatic substitutions for the chosen `gameweek`(s).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{User$get_automatic_substitutions(gameweek = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{gameweek}}{`numeric(n)`. The gameweek history to return (default: `NULL` returns all gameweeks).}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A vector of active chips.
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{user$get_automatic_substitutions()
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_user_history"></a>}}
\if{latex}{\out{\hypertarget{method-get_user_history}{}}}
\subsection{Method \code{get_user_history()}}{
Get a `list` of the user's `gameweek` history.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{User$get_user_history(gameweek = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{gameweek}}{`numeric(n)`. The gameweek history to return (default: `NULL` returns all gameweeks).}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A `list` of `data.frame`s of the user's gameweek history.
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{user$get_user_history()
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_transfers"></a>}}
\if{latex}{\out{\hypertarget{method-get_transfers}{}}}
\subsection{Method \code{get_transfers()}}{
Get a `list` of all the user's transfers, or a `list` of transfers made in the given `gameweek`(s).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{User$get_transfers(gameweek = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{gameweek}}{`numeric(n)`. The gameweek history to return (default: `NULL` returns all gameweeks).}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A `list` of the user's transfers.
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{user$get_transfers()
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-print"></a>}}
\if{latex}{\out{\hypertarget{method-print}{}}}
\subsection{Method \code{print()}}{
Pretty printing of the `User`.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{User$print(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{Not used.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{User$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
