% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{valid_gameweek}
\alias{valid_gameweek}
\title{Determine if the chosen gameweek(s) are valid}
\usage{
valid_gameweek(gameweek)
}
\value{
`TRUE`, invisibly, if the gameweek is valid, otherwise errors.
}
\description{
param gameweek `numeric(n)`. The gameweek(s).
}
\examples{
valid_gameweek(1)
}
