% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R, R/panning2.R
\name{fpl}
\alias{fpl}
\title{Fast Panning for logistic}
\usage{
fpl(X, y, var_mat, seed, ncores, K = 10L, M = 10L)
}
\arguments{
\item{X}{a n x p matrix of regressor}

\item{y}{a n-vector of response}

\item{var_mat}{a matrix of indices for subsetting X}

\item{seed}{an integer for setting the seed (reproducibility)}

\item{ncores}{number of cores for parallel computing (OpenMP)}

\item{K}{number of splits; 10 by default}

\item{M}{number of repetitions; 10 by default}
}
\description{
C++ extension to panning
}
\author{
Samuel Orso
}
