% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/web.R
\name{get_fixture_data}
\alias{get_fixture_data}
\title{Get the Premier League fixture data}
\usage{
get_fixture_data()
}
\value{
Fixture list
}
\description{
Get the Premier League fixture data
}
\examples{
fixture_list <- get_fixture_data()

}
